requirements.txt
python-twitter
facebook
python-dotenv